# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: THE HOMIES ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: paymentMode
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale (normal or troc)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}/cancel:
    post:
      operationId: cancelSale
      tags: [sales]
      summary: Cancel a sale (creates reversal movement)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancellationInput"
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"

  # PARTNERS
...
  /sellers/{id}/sales:
    get:
      operationId: getSellerSales
      tags: [sellers]
      summary: Get sales for a seller (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sales for this seller
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"

  # SEARCH
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, fullName, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, secretary]
        createdAt:
          type: string
          format: date-time
...
    Product:
      type: object
      required: [id, productId, product, status, entryDate, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: string
          description: "Auto-generated like TH001"
        imei:
          type: ["string", "null"]
        product:
          type: string
        brand:
          type: ["string", "null"]
        capacity:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        purchasePrice:
          type: ["number", "null"]
          description: "Hidden from secretary"
        sellingPrice:
          type: ["number", "null"]
        profit:
          type: ["number", "null"]
          description: "Calculated: sellingPrice - purchasePrice"
        status:
          type: string
          enum: [en_stock, chez_partenaire, vendu]
        entryDate:
          type: string
          format: date
        saleDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
...
    Sale:
      type: object
      required: [id, productId, saleType, paymentMode, amount, saleDate, saleTime, sellerId, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        saleType:
          type: string
          enum: [normal, troc, fast_deal]
        paymentMode:
          type: string
          enum: [OM, MOMO, Cash]
        amount:
          type: number
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        clientPhone:
          type: ["string", "null"]
        sellerId:
          type: integer
        seller:
          $ref: "#/components/schemas/User"
        vendorId:
          type: ["integer", "null"]
        vendorName:
          type: ["string", "null"]
        saleDate:
          type: string
          format: date
        saleTime:
          type: string
        cancelled:
          type: boolean
        cancellationReason:
          type: ["string", "null"]
        trocProductId:
          type: ["integer", "null"]
          description: "ID of product received in troc"
        createdAt:
          type: string
          format: date-time
    SaleInput:
      type: object
      required: [productId, saleType, paymentMode, amount]
      properties:
        productId:
          type: integer
        saleType:
          type: string
          enum: [normal, troc, fast_deal]
        paymentMode:
          type: string
          enum: [OM, MOMO, Cash]
        amount:
          type: number
        clientName:
          type: string
        clientPhone:
          type: string
        vendorId:
          type: integer
        vendorName:
          type: string
        trocImei:
          type: string
        trocProduct:
          type: string
        trocBrand:
          type: string
        trocCapacity:
          type: string
        trocColor:
          type: string
    CancellationInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string

    # PARTNER
...